added proper exit methods to test scripts

+++ test/f2s_s2f_integrity_test.sh
@@ -10,6 +10,13 @@ NC='\033[0m' # No Color
 RED='\033[0;31m'
 GREEN='\033[0;32m'
 
+# terminate script
+die() {
+    echo -e "${RED}$1${NC}" >&2
+    echo
+    exit 1
+}
+
 if [ "$1" = 'mem' ]; then
     SLOW5_EXEC="valgrind --leak-check=full --error-exitcode=1 $REL_PATH/../slow5tools"
 else
@@ -19,7 +26,7 @@ fi
 FAST5_DIR="$REL_PATH/data/raw/f2s_s2f_integrity/"
 TEMP_DIR="$REL_PATH/data/out/f2s_s2f_integrity/"
 if [[ "$#" -ge 2 ]]; then
-	FAST5_DIR=$1
+  FAST5_DIR=$1
   TEMP_DIR="$2/f2s_s2f_integrity_test"
 fi
 
@@ -32,58 +39,46 @@ if [[ "$#" -eq 4 ]]; then
     SLOW5_FORMAT=$4
 fi
 
+test -d $TEMP_DIR && rm -r "$TEMP_DIR"
+mkdir "$TEMP_DIR" || die "Creating $TEST_DIR failed"
+mkdir "$F2S_atm1_OUTPUT" || die "Creating $F2S_atm1_OUTPUT failed"
+mkdir "$S2F_OUTPUT" || die "Creating $S2F_OUTPUT failed"
+mkdir "$F2S_atm2_OUTPUT" || die "Creating $F2S_atm2_OUTPUT failed"
 
-rm -r "$TEMP_DIR"
-mkdir "$TEMP_DIR" || exit 1
-mkdir "$F2S_atm1_OUTPUT" || exit 1
-mkdir "$S2F_OUTPUT" || exit 1
-mkdir "$F2S_atm2_OUTPUT" || exit 1
 
 if [[ $* == *-f* ]];then
   clean_fscache
 fi
 echo "-------------------f2s attempt 1-------------------"
 echo
-if ! $SLOW5_EXEC f2s $FAST5_DIR -d $F2S_atm1_OUTPUT --iop 64 $SLOW5_FORMAT; then
-    echo "f2s attempt 1 failed" 
-    exit 1
-fi
-
+$SLOW5_EXEC f2s $FAST5_DIR -d $F2S_atm1_OUTPUT --iop 64 $SLOW5_FORMAT || die "f2s attempt 1 failed"
 if [[ $* == *-f* ]];then
   clean_fscache
 fi
 echo
 echo "-------------------s2f attempt-------------------"
 echo
-if ! $SLOW5_EXEC s2f $F2S_atm1_OUTPUT -d $S2F_OUTPUT --iop 64; then
-    echo "s2f failed"
-    exit 1
-fi
-
+$SLOW5_EXEC s2f $F2S_atm1_OUTPUT -d $S2F_OUTPUT --iop 64 || die "s2f failed"
 if [[ $* == *-f* ]];then
   clean_fscache
 fi
 echo
 echo "-------------------f2s attempt 2-------------------"
 echo
-if ! $SLOW5_EXEC f2s $S2F_OUTPUT -d $F2S_atm2_OUTPUT --iop 64 $SLOW5_FORMAT; then
-    echo "f2s attempt 2 failed"
-    exit 1
-fi
-
+$SLOW5_EXEC f2s $S2F_OUTPUT -d $F2S_atm2_OUTPUT --iop 64 $SLOW5_FORMAT || die "f2s attempt 2 failed"
 echo "running diff on f2s attempt 1 and f2s attempt 2"
 diff -s $F2S_atm1_OUTPUT $F2S_atm2_OUTPUT &>/dev/null
 
 if [ $? -eq 0 ]; then
-	echo -e "${GREEN}SUCCESS: f2s and s2f conversions are consistent!${NC}"
+    echo -e "${GREEN}SUCCESS: f2s and s2f conversions are consistent!${NC}"
 elif [ $? -eq 1 ]; then
-	echo -e "${RED}FAILURE: f2s and s2f conversions are not consistent${NC}"
-	exit 1
+    echo -e "${RED}FAILURE: f2s and s2f conversions are not consistent${NC}"
+    exit 1
 else
-	echo -e "${RED}ERROR: diff failed for some weird reason${NC}"
-	exit 1
+    echo -e "${RED}ERROR: diff failed for some weird reason${NC}"
+    exit 1
 fi
 
+rm -r "$TEMP_DIR" || die "Removing $TEMP_DIR failed"
 
-rm -r "$TEMP_DIR"
-exit
+exit 0
